Moving Learning Items to core
